Did a basic implementation for the board class
The board device is the one that holds the state machine which knows
when the device is running/paused/stopped. It also holds a timer.

The timer calls an ExecutionTask (Name to be changed) that keeps a track
of the execution time and notifies the the board to do the execution if
the board state is running.  This way we keep track of the execution
time separately and when the execute task comes (since all the commands
are ordered temporally it should be able to fire the send command
sequentially.